# docs/openAPI_document/streamflix_openapi.yaml
openapi: 3.1.0
info:
  title: StreamFlix API
  version: v1
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearerAuth: []
...
paths:
...
  /subscriptions:
    post:
      tags:
        - subscription-management
      summary: Create subscription
      description: Create a paid subscription (converts trial if active). Supports JSON, XML, CSV.
      operationId: createSubscription
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
        required: true
      responses:
        '201':
          description: Subscription created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
            text/csv:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '400':
          description: Invalid input - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already has an active subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
...
  /trials:
    get:
      tags:
        - trial-management
      summary: Get trial status
      description: Get trial information for current user. Supports JSON, XML, CSV.
      operationId: getTrial
      responses:
        '200':
          description: Trial information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrialResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/TrialResponse'
            text/csv:
              schema:
                $ref: '#/components/schemas/TrialResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No trial found for current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
  /trials/active:
    get:
      tags:
        - trial-management
      summary: Check if trial is active
      description: Check if user has active trial. Supports JSON, XML, CSV.
      operationId: hasActiveTrial
      responses:
        '200':
          description: Trial status returned
          content:
            application/json:
              schema:
                type: boolean
            application/xml:
              schema:
                type: boolean
            text/csv:
              schema:
                type: boolean
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
...
  /subscriptions/history:
    get:
      tags:
        - subscription-management
      summary: Get subscription history
      description: Get all subscriptions for current user. Supports JSON, XML, CSV.
      operationId: getSubscriptionHistory
      responses:
        '200':
          description: Subscription history retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubscriptionResponse'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubscriptionResponse'
            text/csv:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubscriptionResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
  /subscriptions/current:
    get:
      tags:
        - subscription-management
      summary: Get current subscription
      description: Get active subscription for current user. Supports JSON, XML, CSV.
      operationId: getCurrentSubscription
      responses:
        '200':
          description: Active subscription retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
            text/csv:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No active subscription found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
...
  /subscriptions/cancel:
    delete:
      tags:
        - subscription-management
      summary: Cancel subscription
      description: Cancel current subscription. Supports JSON, XML, CSV.
      operationId: cancelSubscription
      responses:
        '200':
          description: Subscription cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
            text/csv:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No active subscription found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
components:
  schemas:
    ErrorResponse:
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string
...
    CreateSubscriptionRequest:
      properties:
        plan:
          type: string
          enum:
            - SD
            - HD
            - UHD
      required:
        - plan
    SubscriptionResponse:
      properties:
        subscriptionId:
          type: integer
          format: int64
        userId:
          type: string
        plan:
          type: string
        status:
          type: string
        totalPrice:
          type: number
        discountPercentage:
          type: number
        discountedPrice:
          type: number
        discountEndDate:
          type: string
          format: date
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        autoRenew:
          type: boolean
        daysRemaining:
          type: integer
          format: int32
...
    TrialResponse:
      properties:
        trialId:
          type: integer
          format: int64
        userId:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          type: string
        daysRemaining:
          type: integer
          format: int32
        convertedToPaid:
          type: boolean
        convertedDate:
          type: string
          format: date
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT